verify_imports.py
import sys
try:
    import flask
    import pandas
    import simplejson
    import autogen
    print("All packages imported successfully.")
except ImportError as e:
    print(f"Error importing packages: {e}", file=sys.stderr)
    sys.exit(1)
