requirements.txt
Flask
pandas
simplejson
pyautogen
